Remove "Future Hospitalization Projections" chart from location page (but leave most of the code around).
